# Conflicts:
#	include/etl/algorithm.h
#	include/etl/expected.h
#	include/etl/functional.h
#	include/etl/invoke.h
#	include/etl/iterator.h
#	include/etl/memory.h
#	include/etl/private/ranges_mini_variant.h
#	include/etl/private/variant_legacy.h
#	include/etl/ranges.h
#	test/test_algorithm.cpp
#	test/test_functional.cpp
#	test/test_invoke.cpp
#	test/test_iterator.cpp
#	test/test_memory.cpp
#	test/test_ranges.cpp
#	test/vs2022/etl.vcxproj.filters

+++ include/etl/private/chrono/month.h
@@ -38,13 +38,13 @@ namespace etl
   {
     class month;
 
-    ETL_CONSTEXPR14 etl::chrono::month operator +(const etl::chrono::month& m,   const etl::chrono::months& ms) ETL_NOEXCEPT;
-    ETL_CONSTEXPR14 etl::chrono::month operator +(const etl::chrono::months& ms, const etl::chrono::month& m)   ETL_NOEXCEPT;
-    ETL_CONSTEXPR14 etl::chrono::month operator -(const etl::chrono::month& m,   const etl::chrono::months& ms) ETL_NOEXCEPT;
+    ETL_CONSTEXPR14 etl::chrono::month operator+(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT;
+    ETL_CONSTEXPR14 etl::chrono::month operator+(const etl::chrono::months& ms, const etl::chrono::month& m) ETL_NOEXCEPT;
+    ETL_CONSTEXPR14 etl::chrono::month operator-(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT;
 
     namespace private_chrono
     {
-      static ETL_CONSTANT unsigned char days_in_month[13] = { 0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
+      static ETL_CONSTANT unsigned char days_in_month[13] = {0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
     }
 
     //***********************************************************************
@@ -75,7 +75,7 @@ namespace etl
       //***********************************************************************
       /// Pre-increment operator
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month& operator ++() ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month& operator++() ETL_NOEXCEPT
       {
         *this += etl::chrono::months(1);
 
@@ -85,10 +85,10 @@ namespace etl
       //***********************************************************************
       /// Post-increment operator
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month operator ++(int) ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month operator++(int) ETL_NOEXCEPT
       {
         etl::chrono::month temp = *this;
-        
+
         ++*this;
 
         return temp;
@@ -97,7 +97,7 @@ namespace etl
       //***********************************************************************
       /// Pre-decrement operator
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month& operator --() ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month& operator--() ETL_NOEXCEPT
       {
         *this -= etl::chrono::months(1);
 
@@ -107,7 +107,7 @@ namespace etl
       //***********************************************************************
       /// Post-decrement operator
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month operator --(int) ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month operator--(int) ETL_NOEXCEPT
       {
         etl::chrono::month temp = *this;
 
@@ -119,7 +119,7 @@ namespace etl
       //***********************************************************************
       /// Plus-equals operator adding etl::chrono::months
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month& operator +=(const etl::chrono::months& ms) ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month& operator+=(const etl::chrono::months& ms) ETL_NOEXCEPT
       {
         *this = *this + ms;
 
@@ -129,7 +129,7 @@ namespace etl
       //***********************************************************************
       /// Minus-equals operator subtracting etl::chrono::months
       //***********************************************************************
-      ETL_CONSTEXPR14 etl::chrono::month& operator -=(const etl::chrono::months& ms) ETL_NOEXCEPT
+      ETL_CONSTEXPR14 etl::chrono::month& operator-=(const etl::chrono::months& ms) ETL_NOEXCEPT
       {
         *this = *this - ms;
 
@@ -139,8 +139,7 @@ namespace etl
       //***********************************************************************
       /// Returns <b>true</b> if the month is within the valid 1 to 12 range
       //***********************************************************************
-      ETL_NODISCARD
-      ETL_CONSTEXPR14 bool ok() const ETL_NOEXCEPT
+      ETL_NODISCARD ETL_CONSTEXPR14 bool ok() const ETL_NOEXCEPT
       {
         return (value >= 1U) && (value <= 12U);
       }
@@ -151,11 +150,12 @@ namespace etl
       /// else if month > other, returns 1
       /// else returns 0
       //***********************************************************************
-      ETL_NODISCARD
-      ETL_CONSTEXPR14 int compare(const month& other) const ETL_NOEXCEPT 
+      ETL_NODISCARD ETL_CONSTEXPR14 int compare(const month& other) const ETL_NOEXCEPT
       {
-        if (value < other.value) return -1;
-        if (value > other.value) return 1;
+        if (value < other.value)
+          return -1;
+        if (value > other.value)
+          return 1;
 
         return 0;
       }
@@ -176,7 +176,7 @@ namespace etl
     //***********************************************************************
     /// Equality operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator ==(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator==(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) == static_cast<unsigned>(d2));
     }
@@ -184,7 +184,7 @@ namespace etl
     //***********************************************************************
     /// Inequality operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator !=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator!=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return !(d1 == d2);
     }
@@ -192,7 +192,7 @@ namespace etl
     //***********************************************************************
     /// Less-than operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator <(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator<(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) < static_cast<unsigned>(d2));
     }
@@ -200,7 +200,7 @@ namespace etl
     //***********************************************************************
     /// Less-than-or-equal operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator <=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator<=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) <= static_cast<unsigned>(d2));
     }
@@ -208,7 +208,7 @@ namespace etl
     //***********************************************************************
     /// Greater-than operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator >(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator>(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) > static_cast<unsigned>(d2));
     }
@@ -216,7 +216,7 @@ namespace etl
     //***********************************************************************
     /// Greater-than-or-equal operator
     //***********************************************************************
-    inline ETL_CONSTEXPR14 bool operator >=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 bool operator>=(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) >= static_cast<unsigned>(d2));
     }
@@ -225,7 +225,8 @@ namespace etl
     /// Spaceship operator
     //***********************************************************************
 #if ETL_USING_CPP20
-    [[nodiscard]] inline constexpr auto operator <=>(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
+    [[nodiscard]]
+    inline constexpr auto operator<=>(const etl::chrono::month& d1, const etl::chrono::month& d2) ETL_NOEXCEPT
     {
       return (static_cast<unsigned>(d1) <=> static_cast<unsigned>(d2));
     }
@@ -235,7 +236,7 @@ namespace etl
     /// Add etl::chrono::months to etl::chrono::month
     ///\return etl::chrono::month
     //***********************************************************************
-    inline ETL_CONSTEXPR14 etl::chrono::month operator +(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month operator+(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT
     {
       unsigned int value = static_cast<unsigned int>(m);
 
@@ -250,7 +251,7 @@ namespace etl
 
       // Adjust to allow a limited +-11 month delta
       value += 11U;
-      value += delta;
+      value += static_cast<unsigned int>(delta);
       value %= 12U;
       ++value;
 
@@ -261,7 +262,7 @@ namespace etl
     /// Add etl::chrono::month to etl::chrono::months
     ///\return etl::chrono::month
     //***********************************************************************
-    inline ETL_CONSTEXPR14 etl::chrono::month operator +(const etl::chrono::months& ms, const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month operator+(const etl::chrono::months& ms, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return m + ms;
     }
@@ -270,7 +271,7 @@ namespace etl
     /// Subtract etl::chrono::months from etl::chrono::month
     ///\return etl::chrono::month
     //***********************************************************************
-    inline ETL_CONSTEXPR14 etl::chrono::month operator -(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month operator-(const etl::chrono::month& m, const etl::chrono::months& ms) ETL_NOEXCEPT
     {
       return m + etl::chrono::months(-ms.count());
     }
@@ -279,7 +280,7 @@ namespace etl
     /// Subtract etl::chrono::month from etl::chrono::month
     ///\return etl::chrono::months
     //***********************************************************************
-    inline ETL_CONSTEXPR14 etl::chrono::months operator -(const etl::chrono::month& m1, const etl::chrono::month& m2) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::months operator-(const etl::chrono::month& m1, const etl::chrono::month& m2) ETL_NOEXCEPT
     {
       if (m1.ok() && m2.ok())
       {
@@ -303,33 +304,33 @@ namespace etl
     }
 
 #if ETL_USING_CPP17
-    inline constexpr etl::chrono::month January{ 1 };
-    inline constexpr etl::chrono::month February{ 2 };
-    inline constexpr etl::chrono::month March{ 3 };
-    inline constexpr etl::chrono::month April{ 4 };
-    inline constexpr etl::chrono::month May{ 5 };
-    inline constexpr etl::chrono::month June{ 6 };
-    inline constexpr etl::chrono::month July{ 7 };
-    inline constexpr etl::chrono::month August{ 8 };
-    inline constexpr etl::chrono::month September{ 9 };
-    inline constexpr etl::chrono::month October{ 10 };
-    inline constexpr etl::chrono::month November{ 11 };
-    inline constexpr etl::chrono::month December{ 12 };
-#else     
-    static ETL_CONSTANT etl::chrono::month January{ 1 };
-    static ETL_CONSTANT etl::chrono::month February{ 2 };
-    static ETL_CONSTANT etl::chrono::month March{ 3 };
-    static ETL_CONSTANT etl::chrono::month April{ 4 };
-    static ETL_CONSTANT etl::chrono::month May{ 5 };
-    static ETL_CONSTANT etl::chrono::month June{ 6 };
-    static ETL_CONSTANT etl::chrono::month July{ 7 };
-    static ETL_CONSTANT etl::chrono::month August{ 8 };
-    static ETL_CONSTANT etl::chrono::month September{ 9 };
-    static ETL_CONSTANT etl::chrono::month October{ 10 };
-    static ETL_CONSTANT etl::chrono::month November{ 11 };
-    static ETL_CONSTANT etl::chrono::month December{ 12 };
+    inline constexpr etl::chrono::month January{1};
+    inline constexpr etl::chrono::month February{2};
+    inline constexpr etl::chrono::month March{3};
+    inline constexpr etl::chrono::month April{4};
+    inline constexpr etl::chrono::month May{5};
+    inline constexpr etl::chrono::month June{6};
+    inline constexpr etl::chrono::month July{7};
+    inline constexpr etl::chrono::month August{8};
+    inline constexpr etl::chrono::month September{9};
+    inline constexpr etl::chrono::month October{10};
+    inline constexpr etl::chrono::month November{11};
+    inline constexpr etl::chrono::month December{12};
+#else
+    static ETL_CONSTANT etl::chrono::month January{1};
+    static ETL_CONSTANT etl::chrono::month February{2};
+    static ETL_CONSTANT etl::chrono::month March{3};
+    static ETL_CONSTANT etl::chrono::month April{4};
+    static ETL_CONSTANT etl::chrono::month May{5};
+    static ETL_CONSTANT etl::chrono::month June{6};
+    static ETL_CONSTANT etl::chrono::month July{7};
+    static ETL_CONSTANT etl::chrono::month August{8};
+    static ETL_CONSTANT etl::chrono::month September{9};
+    static ETL_CONSTANT etl::chrono::month October{10};
+    static ETL_CONSTANT etl::chrono::month November{11};
+    static ETL_CONSTANT etl::chrono::month December{12};
 #endif
-  }
+  } // namespace chrono
 
   //*************************************************************************
   /// Hash function for etl::chrono::month
@@ -341,12 +342,11 @@ namespace etl
     size_t operator()(const etl::chrono::month& m) const
     {
       etl::chrono::month::rep value = static_cast<etl::chrono::month::rep>(static_cast<unsigned>(m));
-      const uint8_t* p = reinterpret_cast<const uint8_t*>(&value);
+      const uint8_t*          p     = reinterpret_cast<const uint8_t*>(&value);
 
       return etl::private_hash::generic_hash<size_t>(p, p + sizeof(value));
     }
   };
 #endif
 
-
-}
+} // namespace etl
