+++ test/test_crc16.cpp
@@ -29,9 +29,9 @@ SOFTWARE.
 #include "unit_test_framework.h"
 
 #include <iterator>
+#include <stdint.h>
 #include <string>
 #include <vector>
-#include <stdint.h>
 
 #include "etl/crc16.h"
 #include "etl/endianness.h"
@@ -65,7 +65,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint16_t crc = crc_calculator;
@@ -104,17 +104,17 @@ namespace
     //*************************************************************************
     TEST(test_crc16_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint16_t crc1 = etl::crc16(data1.begin(), data1.end());
       uint16_t crc2 = etl::crc16((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -140,8 +140,8 @@ namespace
     //*************************************************************************
     TEST(test_crc16_16_constexpr)
     {
-      constexpr char data[] = "123456789";
-      constexpr uint16_t crc = etl::crc16_t16(data, data + 9);
+      constexpr char     data[] = "123456789";
+      constexpr uint16_t crc    = etl::crc16_t16(data, data + 9);
 
       CHECK_EQUAL(0xBB3DU, crc);
     }
@@ -156,7 +156,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint16_t crc = crc_calculator;
@@ -195,17 +195,17 @@ namespace
     //*************************************************************************
     TEST(test_crc16_16_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint16_t crc1 = etl::crc16_t16(data1.begin(), data1.end());
       uint16_t crc2 = etl::crc16_t16((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -231,8 +231,8 @@ namespace
     //*************************************************************************
     TEST(test_crc16_4_constexpr)
     {
-      constexpr char data[] = "123456789";
-      constexpr uint16_t crc = etl::crc16_t4(data, data + 9);
+      constexpr char     data[] = "123456789";
+      constexpr uint16_t crc    = etl::crc16_t4(data, data + 9);
 
       CHECK_EQUAL(0xBB3DU, crc);
     }
@@ -247,7 +247,7 @@ namespace
 
       for (size_t i = 0UL; i < data.size(); ++i)
       {
-        crc_calculator.add(data[i]);
+        crc_calculator.add(static_cast<uint8_t>(data[i]));
       }
 
       uint16_t crc = crc_calculator;
@@ -286,17 +286,17 @@ namespace
     //*************************************************************************
     TEST(test_crc16_4_add_range_endian)
     {
-      std::vector<uint8_t>  data1 = { 0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U };
+      std::vector<uint8_t>  data1 = {0x01U, 0x02U, 0x03U, 0x04U, 0x05U, 0x06U, 0x07U, 0x08U};
       std::vector<uint32_t> data2;
       if (etl::endianness::value() == etl::endian::little)
       {
-        data2 = { 0x04030201UL, 0x08070605UL };
+        data2 = {0x04030201UL, 0x08070605UL};
       }
       else
       {
-        data2 = { 0x01020304UL, 0x05060708UL };
+        data2 = {0x01020304UL, 0x05060708UL};
       }
-      std::vector<uint8_t>  data3 = { 0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U };
+      std::vector<uint8_t> data3 = {0x08U, 0x07U, 0x06U, 0x05U, 0x04U, 0x03U, 0x02U, 0x01U};
 
       uint16_t crc1 = etl::crc16_t4(data1.begin(), data1.end());
       uint16_t crc2 = etl::crc16_t4((uint8_t*)&data2[0], (uint8_t*)(&data2[0] + data2.size()));
@@ -306,5 +306,4 @@ namespace
       CHECK_EQUAL(crc1, crc3);
     }
   }
-}
-
+} // namespace
